Edited file systems for modules
